Rename files after the service
