+++ packages/bridge-ui/src/App.svelte
@@ -42,13 +42,13 @@
   import { ethers } from 'ethers';
   import type { Prover } from './domain/proof';
   import { successToast } from './utils/toast';
-  import { StorageService } from './storage/service';
+  import { StorageService } from './storage/StorageService';
   import { MessageStatus } from './domain/message';
   import BridgeABI from './constants/abi/Bridge';
   import { providers } from './store/providers';
   import HeaderAnnouncement from './components/HeaderAnnouncement.svelte';
   import type { TokenService } from './domain/token';
-  import { CustomTokenService } from './storage/customTokenService';
+  import { CustomTokenService } from './storage/CustomTokenService';
   import { userTokens, tokenService } from './store/userToken';
   import RelayerAPIService from './relayer-api/service';
   import type { RelayerAPI } from './domain/relayerApi';
@@ -145,7 +145,6 @@
       const userAddress = await store.getAddress();
 
       const apiTxs = await $relayerApi.GetAllByAddress(userAddress);
-
 
       const blockInfoMap = await $relayerApi.GetBlockInfo();
       relayerBlockInfoMap.set(blockInfoMap);
